Update ffi to llhttp 6.0.5

+++ ffi/ext/llhttp/llhttp.c
@@ -1126,7 +1126,7 @@ static llparse_state_t llhttp__internal__run(
     case s_n_llhttp__internal__n_consume_content_length:
     s_n_llhttp__internal__n_consume_content_length: {
       size_t avail;
-      size_t need;
+      uint64_t need;
 
       avail = endp - p;
       need = state->content_length;
@@ -1481,7 +1481,7 @@ static llparse_state_t llhttp__internal__run(
     case s_n_llhttp__internal__n_consume_content_length_1:
     s_n_llhttp__internal__n_consume_content_length_1: {
       size_t avail;
-      size_t need;
+      uint64_t need;
 
       avail = endp - p;
       need = state->content_length;
@@ -8700,7 +8700,7 @@ static llparse_state_t llhttp__internal__run(
     case s_n_llhttp__internal__n_consume_content_length:
     s_n_llhttp__internal__n_consume_content_length: {
       size_t avail;
-      size_t need;
+      uint64_t need;
 
       avail = endp - p;
       need = state->content_length;
@@ -9048,7 +9048,7 @@ static llparse_state_t llhttp__internal__run(
     case s_n_llhttp__internal__n_consume_content_length_1:
     s_n_llhttp__internal__n_consume_content_length_1: {
       size_t avail;
-      size_t need;
+      uint64_t need;
 
       avail = endp - p;
       need = state->content_length;
